+++ docs_src/tutorial/init_and_define_model/tutorial000.py
@@ -8,5 +8,5 @@ class Player(Document):
     country_code: str
     rating: Optional[int] = None
 
-    class Config(Document.Config):
+    class ODMConfig(Document.ODMConfig):
         collection_name = "player"