+++ docs_src/tutorial/relation/many_to_many_relationship/tutorial001.py
@@ -34,7 +34,7 @@ class Player(Document):
 
     country: Optional[Country] = Relationship(local_field="country_id")
 
-    class Config(Document.Config):
+    class ODMConfig(Document.ODMConfig):
         indexes = [
             IndexModel([("country_id", ASCENDING)]),
         ]