+++ mongodb_odm/connection.py
@@ -90,6 +90,10 @@ def get_client() -> MongoClient[Any]:
         return __connection_obj.client
 
 
+def drop_database(database: Optional[str] = None) -> None:
+    db(database).command("dropDatabase")
+
+
 def db(database: Optional[str] = None) -> Database[Any]:
     if database:
         """