+++ mongodb_odm/types.py
@@ -1,6 +1,7 @@
 from typing import Any, Dict, Mapping, Sequence, Tuple, Union
 
 from bson import ObjectId
+from pydantic_core import core_schema
 
 # Common types
 DICT_TYPE = Dict[str, Any]
@@ -8,24 +9,36 @@ SORT_TYPE = Union[str, Sequence[Tuple[str, Union[int, str, Mapping[str, Any]]]]]
 
 
 class ODMObjectId(ObjectId):
-    """
-    Since ObjectId has not validation and raise error as pydantic type
-    we extend ObjectId for type support.
-    """
-
     @classmethod
-    def __get_validators__(cls: Any) -> Any:
-        yield cls.validate
+    def __get_pydantic_core_schema__(
+        cls, _source_type: Any, _handler: Any
+    ) -> core_schema.CoreSchema:
+        object_id_schema = core_schema.chain_schema(
+            [
+                core_schema.str_schema(),
+                core_schema.no_info_plain_validator_function(cls.validate),
+            ]
+        )
+        return core_schema.json_or_python_schema(
+            json_schema=object_id_schema,
+            python_schema=core_schema.union_schema(
+                [core_schema.is_instance_schema(ObjectId), object_id_schema]
+            ),
+            serialization=core_schema.plain_serializer_function_ser_schema(
+                lambda x: ObjectId(x)
+            ),
+        )
 
     @classmethod
     def validate(cls, v: Union[str, ObjectId]) -> ObjectId:
         if isinstance(v, ObjectId):
             """No conversion needed if type are already ObjectId"""
             return v
-        elif isinstance(v, str):
+        try:
             """
             If value is valid string for ObjectId then convert it ObjectId
             or it will raise error from ObjectId validation
             """
             return ObjectId(v)
-        raise TypeError("Invalid data. ObjectId required")
+        except Exception as e:
+            raise ValueError("Invalid data. ObjectId required") from e