+++ mongodb_odm/utils/utils.py
@@ -1,11 +1,15 @@
 import re
-from typing import Any, Dict, List, Optional, Type
+import types
+from typing import Any, Dict, List, Optional, Type, Union
 
 from pydantic import BaseModel
+from typing_extensions import get_args, get_origin
 
-from ..fields import _RelationshipInfo
+from ..fields import RelationshipInfo
 from ._internal_models import RelationalFieldInfo
 
+UnionType = getattr(types, "UnionType", Union)
+NoneType = type(None)
 pattern = re.compile(r"(?<!^)(?=[A-Z])")
 
 
@@ -15,19 +19,19 @@ def camel_to_snake(string: str) -> str:
 
 def get_database_name(model: Any) -> Optional[Any]:
     """
-    Get the database name if the model Config has a user-defined database name.
+    Get the database name if the model ODMConfig has a user-defined database name.
 
     model: Document type
     """
-    if hasattr(model.Config, "database"):
-        return model.Config.database
+    if hasattr(model.ODMConfig, "database"):
+        return model.ODMConfig.database
     return None
 
 
 def convert_model_to_collection(model: Any) -> str:
     """
     Get the collection name from the model.
-    Users could define collection names in model Config.
+    Users could define collection names in model ODMConfig.
     Otherwise, we will convert the model name to a snake case collection name.
 
     The user defines collection_name as a higher priority here.
@@ -35,13 +39,40 @@ def convert_model_to_collection(model: Any) -> str:
     model: Document type
     """
     if (
-        hasattr(model.Config, "collection_name")
-        and model.Config.collection_name is not None
+        hasattr(model.ODMConfig, "collection_name")
+        and model.ODMConfig.collection_name is not None
     ):
-        """By default model has Config in BaseModel"""
-        return str(model.Config.collection_name)
+        """By default model has ODMConfig in BaseModel"""
+        return str(model.ODMConfig.collection_name)
     else:
         return camel_to_snake(model.__name__)
+
+
+def _is_union_type(t: Any) -> bool:
+    return t is UnionType or t is Union
+
+
+def get_type_from_field(field: Any) -> Any:
+    type_: Any = field.annotation
+    # Resolve Optional fields
+    if type_ is None:
+        raise ValueError("Missing field type")
+    origin = get_origin(type_)
+    if origin is None:
+        return type_
+    if _is_union_type(origin):
+        bases = get_args(type_)
+        if len(bases) > 2:
+            raise ValueError("Cannot have a (non-optional) union as a ODM field")
+        # Non optional unions are not allowed
+        if bases[0] is not NoneType and bases[1] is not NoneType:
+            raise ValueError("Cannot have a (non-optional) union as a ODM field")
+        # Optional unions are allowed
+        return bases[0] if bases[0] is not NoneType else bases[1]
+    elif origin is list:
+        inner_type_ = get_args(type_)[0]
+        return inner_type_
+    return origin
 
 
 def _get_fields_info(
@@ -52,11 +83,11 @@ def _get_fields_info(
     """
     field_data: Dict[str, RelationalFieldInfo] = {}
     for field in fields:
-        obj = cls.__fields__[field]
-        if obj.default.local_field not in cls.__fields__:
+        field_type_obj = cls.model_fields[field]
+        if field_type_obj.default.local_field not in cls.model_fields:
             # Check Relationship local_field exists in the model
             raise Exception(
-                f'Invalid field "{obj.default.local_field}" in Relationship'
+                f'Invalid field "{field_type_obj.default.local_field}" in Relationship'
             )
         """
         Example relationship
@@ -68,10 +99,11 @@ def _get_fields_info(
             author_id: ODMObjectId = Field(...)
             author: Optional[User] = Relationship(local_field="author_id")
         """
+        field_type = get_type_from_field(field_type_obj)
         field_data[field] = RelationalFieldInfo(
-            model=obj.type_,  # User
-            local_field=obj.default.local_field,  # author_id
-            related_field=obj.default.related_field,  # author
+            model=field_type,  # User
+            local_field=field_type_obj.default.local_field,  # author_id
+            related_field=field_type_obj.default.related_field,  # author
         )
     return field_data
 
@@ -85,8 +117,8 @@ def get_relationship_fields_info(
     cls: Document
     """
     fields_name = []
-    for field_name, field_info in cls.__fields__.items():
+    for field_name, field_info in cls.model_fields.items():
         """Get all fields that are related to a specific model."""
-        if type(field_info.default) is _RelationshipInfo:
+        if type(field_info.default) is RelationshipInfo:
             fields_name.append(field_name)
     return _get_fields_info(cls, fields_name)