+++ tests/test_create.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 def test_create_course():
     user = get_user()
     _ = Course(
-        author_id=user._id,
+        author_id=user.id,
         title="Course Title",
     ).create()
 
@@ -18,7 +18,7 @@ def test_create_course():
 def test_inheritance_model_create():
     user = get_user()
     course = Course(
-        author_id=user._id,
+        author_id=user.id,
         title="Course Title",
         short_description="Short Description",
     ).create()
@@ -28,11 +28,11 @@ def test_inheritance_model_create():
     ).create()
 
     content_image = ContentImage(
-        course_id=course._id, image_path="/image/path/image.png"
+        course_id=course.id, image_path="/image/path/image.png"
     ).create()
 
     content_count = Content.count_documents(
-        {"_id": {"$in": [content_description._id, content_image._id]}}
+        {"_id": {"$in": [content_description.id, content_image.id]}}
     )
 
     assert (