+++ docs/release-notes.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## Latest Changes
+
+### Migration
+
+- Add support for Pydantic V2. PR [#22](https://github.com/nayan32biswas/mongodb-odm/pull/22) by [@nayan32biswas](https://github.com/nayan32biswas).
+- Add support for Python 3.12. PR [#22](https://github.com/nayan32biswas/mongodb-odm/pull/22) by [@nayan32biswas](https://github.com/nayan32biswas).
+
+### Breaking Change
+
+- Rename the model inner config class from `Config` to `ODMConfig`. PR [#22](https://github.com/nayan32biswas/mongodb-odm/pull/22) by [@nayan32biswas](https://github.com/nayan32biswas).
+- Now `_id` field was a `PrivateAttr`. It will behave as private attr as Pydantic define. PR [#22](https://github.com/nayan32biswas/mongodb-odm/pull/22) by [@nayan32biswas](https://github.com/nayan32biswas).
+
 ## 0.2.5
 
 ### Fixes